x509: factor `GeneralName` tests under `tests/`
An alternative to conditionally gating `std` linkage on `tests` is
placing such tests under the `tests/` directory.

This commit places the `GeneralName` tests which use `rstest` under the
`tests/` directory.
