+++ x509/src/ext/pkix/name/other.rs
@@ -23,14 +23,11 @@ pub struct OtherName<'a> {
 #[cfg(test)]
 fn test() {
     use alloc::string::ToString;
-    use std::eprintln;
-
     use der::{Decodable, Encodable};
     use hex_literal::hex;
 
     let input = hex!("3021060A2B060104018237140203A0130C1155706E5F323134393530313330406D696C");
     let decoded = OtherName::from_der(&input).unwrap();
-    eprintln!("{:?}", decoded.value);
 
     let onval = decoded.value.utf8_string().unwrap();
     assert_eq!(onval.to_string(), "Upn_214950130@mil");
