+++ x509/src/lib.rs
@@ -11,7 +11,7 @@
 
 extern crate alloc;
 
-#[cfg(any(feature = "std", test))]
+#[cfg(feature = "std")]
 extern crate std;
 
 pub mod anchor;
